Direct mmap entry access for reduced memory usage (#17) (#44)
* Direct mmap entry access for reduced memory usage (#17)

Eliminate []Entry slice allocation by accessing entries directly from mmap'd
file data. This reduces memory usage from ~40MB (1M entries) to near-zero
(kernel-managed mmap pages).

Changes:
- Add RawEntry packed struct (27 bytes) matching on-disk format exactly
- Add getEntry(idx) method for on-demand entry parsing from mmap
- Add getMkvOffset(idx) for efficient binary search without full parsing
- Replace loadEntries() with lightweight initEntryAccess()
- Update findEntriesForRange() to use getEntry() with cache support
- Remove Entries field from File struct
- Update DESIGN.md with new entry access optimization section

The implementation uses byte arrays with explicit little-endian decoding
for portable unaligned access (SourceOffset is at byte 17, not 8-byte aligned).


* Fix gofmt formatting in format_test.go


* Address PR feedback: concurrency, binary search, format v3

- Add cacheMu sync.Mutex for thread-safe cache access
- Update binary search to use getMkvOffset + getEntryLength (16 bytes vs 28)
- Expand Source field from uint8 to uint16 (file format v3, 28-byte entries)
- Add InitEntryAccess() method and error handling in Info()
- Add comprehensive reader_test.go for cache testing
- Update FILE_FORMAT.md documentation for v3

Addresses Copilot review feedback and owner request for >256 source file support.


* Fix gofmt formatting in reader_test.go


* Apply suggestion from @Copilot


* Fix comment inconsistencies after v3 format change

- Fix "not 27" to "not 28" in binary search comment
- Remove duplicate Info() comment
- Update DESIGN.md RawEntry size from 27 to 28 bytes
- Update DESIGN.md version section from v2 to v3


---------

+++ internal/matcher/matcher.go
@@ -23,12 +23,12 @@ const (
 
 // Entry represents a region in the MKV file and where its data comes from.
 type Entry struct {
-	MkvOffset        int64 // Start offset in the MKV file
-	Length           int64 // Length of this region
-	Source           uint8 // 0 = delta, 1+ = source file index + 1
-	SourceOffset     int64 // Offset in source file (or ES offset for ES-based sources)
-	IsVideo          bool  // For ES-based sources: whether this is video or audio data
-	AudioSubStreamID byte  // For ES-based audio: sub-stream ID (0x80-0x87=AC3, etc.)
+	MkvOffset        int64  // Start offset in the MKV file
+	Length           int64  // Length of this region
+	Source           uint16 // 0 = delta, 1+ = source file index + 1 (supports up to 65535 files)
+	SourceOffset     int64  // Offset in source file (or ES offset for ES-based sources)
+	IsVideo          bool   // For ES-based sources: whether this is video or audio data
+	AudioSubStreamID byte   // For ES-based audio: sub-stream ID (0x80-0x87=AC3, etc.)
 }
 
 // Result contains the results of the matching process.
@@ -576,7 +576,7 @@ func (m *Matcher) buildEntries() ([]Entry, []byte) {
 			entries = append(entries, Entry{
 				MkvOffset:        pos,
 				Length:           regionLen,
-				Source:           uint8(inRegion.fileIndex + 1),
+				Source:           uint16(inRegion.fileIndex + 1),
 				SourceOffset:     inRegion.srcOffset + offsetInRegion,
 				IsVideo:          inRegion.isVideo,
 				AudioSubStreamID: inRegion.audioSubStreamID,
